requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
Django==3.1.2
django-crispy-forms==1.9.2
django-filter==2.4.0
django-mathfilters==1.0.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
jeepney==0.4.3
jsonconfig-tool==1.5
keyring==21.4.0
messages==0.5.0
Pillow==7.2.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycparser==2.20
python-box==5.1.1
pytz==2020.1
requests==2.24.0
SecretStorage==3.1.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
validus==0.3.0
whitenoise==5.2.0
zipp==3.2.0
